Mark failing tests as XFAILs instead of disabling them.
GTest has a mechanism for marking tests as disabled when
they're not ready. Automake has another method, where executables
can be marked as 'expected to fail'.

The gtest method has finer granularity, but the automake
method is more transparent about what's working and what's
not.

This commit enables everything but
CueIdLanguage.DISABLED_LongStringMultipleLanguages which hangs in
an inifinite loop. All of the unit tests are added to the XFAIL
list since none of them completely pass at this point.