Correctly zero the webvtt_state::v union.
A number of places in the code zero the 'v' union by assigning
zero to v.value. However, v.value is an unsigned int, and therefore
shorter than the pointer values in the union on 64-bit systems.

This leads to non-NULL v.cue or v.text pointers which point to
random areas of memory, since only the lower 32 bits of the pointer
were cleared.

Related to issue #15. With this commit it still segfaults in
find_bytes() but this time because it doesn't null-check its
arguments, not because they point to unallocated memory.