Add ref counting to c nodes.
I've added ref counting to the c nodes. I've added a webvtt_ref_node function
to make it easy to increase the ref and I've also renamed the webvtt_delete_node
function to webvtt_release_node in order to make it more clear what the
function is actually doing. Finally I've made them public API as we probably
want to be using these with the Node bindings. I've gotten rid of the
declarations for webvtt_delete_leaf_node and webvtt_delete_internal_node
because we aren't currently using them.