Fix segfaults in ExtraNewLinesAfterCue.
I've fixed the segfaul in FileStructure.ExtraNewLinesAfterCue.
Problem was that it was trying to parse a cue when it got to the
end of the buffer and there was just new lines after it. I've set it
up to now consume all white space after a cue or cue text has been
parsed. It's seemed to fix a lot of the other tests too.