Move parse_timestamp to be visible in internal API.
I've moved the parse_timestamp function to be visible in the internal API.
This is so the cue text parser can also be able to call it when it parses
a time stamp in the cue text. I've removed the webvtt_parser self argument
as it's currently not being used by the function and this argument doesn't
really make sense from the perspective of the cue text parser.