Add ref counting to Stringlist.
I've added ref count to the string list struct. This will be useful because it will
make it easier to keep track of string lists internally, but also because string lists
are currently used publicly, so that is something that we want to keep track of.

Remove cilanuage_unittest from the known valgrind failures list.