Fix utf16At string binding function.
For most of the utf16At calls we are passing in an index of offset of 0,
i.e. the utf8 character is at the very beginning of the string. The utf16At
string binding function calls the webvtt_skip_utf8 function which currently
views an offset of 0 as invalid and returns 0. I've changed it so that a 0
offset is valid and will return true.