Change create_node_from_token to accept a null node.
- Node can actually be null and has to be in some instances. It doesn't matter if node is
null because node is the var that we will be loading the new node into. It has to acccept
a null node because the first time this function is called it is null due to never having
been used before. Because it does not accept null nodes it is currently blocking all nodes
from being created.
- I've also removed plitalictag, bold, underline, class, and voice unittests from the XFail
list as they are now all passing (not valgrind though). This is due to the cumulative
changes from my last 10 or so patches.