Fix attach internal node function to resolve memory leaks.
When growing the internal node's children we don't have to worry about
allocating space for the webvtt node pointers, which the current attach
internal node function is doing. All we need to worry about is
allocating space for the new bigger child list and then copy over
the node pointers from the old child list.