Implemented helper routines to assist in cueparameter parsing
I've provided 2 routines (token_in_list and find_token) which search
for or return the index of a token in a null-terminated array of
tokens.

These routines are helpful for providing a general way of parsing a
given sub-grammar in a very general way that can be shared by several
different users (Such as cueparameter parsing for the 5 parameter
kinds)

This code was originally taken from #73, but has been modified and
moved into parser_internal.h so that unit tests of these helper
functions can be written.