Do not #include <windows.h> in parser_internal.h
While public includes should be safe due to our pseudo-namespacing
with the WEBVTT_/webvtt_ prefixes, parser_internal.h could not safely
include <windows.h> as SIZE, amont other things, would be defined by
windows.

mingw-gcc (and mingw cross compilers) support the
__declspec(dllimport) and __stdcall keywords as well as the native gcc
keywords, so it does not appear to hurt to just prototype the function
at the head of parser_internal.h

It could potentially require us to differentiate between windows CE
and different versions of the windows SDK, but for now it is okay.