Fix multiline cue text error.
The problem was that we were ending on any line that ends with a newline
character. Therefore, we would only parse one line of cue text. I've
set it up so that it ends parsing cue text on a line that is either
empty or consists of only a new line character.