Passing all but two 'line' cuesetting tests.
I've integrated the webvtt_parse_line() routine from my large patch,
and am calling it in the same fashion as I am calling the parse_align
routine, so we get some uniform results from the two tests.

There are still two tests failing, which stems from not having a
standard on which errors to return and which errors to expect. These
two other tests will also require unit test changes, and will be
corrected in a followup pull request.