Fixes #281 - Add no-op statement to fix -Wunused warnings (cuetext.c)
This doesn't evaluate to any machine code being generated, it's a
no-op. However, it will prevent the warning from being emitted.

I've left a comment documenting the purpose of the statements, and a
TODO asking 'self' to be used for reporting errors in the future.

Maybe we should define an UNUSED(variable) macro somewhere to make it
semantically clear what the statements are doing, without comments.