Fix filestructure_unittest
Carefully ensure that lines and columns are being adjusted correctly
(except that multibyte characters still occupy multiple columns, will
correct in future).

Renamed missing-cue-identifier.vtt (This file is actually missing
cue-times, not a cue identifier).

Made line/column for incomplete cues relatively expected value (But
honestly I'm not sure what the best way to do this is, and the way it
is now is probably wrong)

Removed the *mode parameter from read_cuetext() and parse_webvtt(),
because this always points to self->mode anyways. Updated
readcuetext_unittest to accomodate this.

Removed duplicate test (MultiCueNoNewlineBetweenCues and
MissingNewLineBetweenCues are testing the same thing, so the latter
was removed)