Simplify cue-settings API
No longer exporting `webvtt_cue_set_setting_word` (`_word` replaced with
`_from_string` to be less nonsensicle). The reason for this is that it is
not something that should be particularly useful to an application (they would
surely use `_set_setting` instead, if they really wanted that sort of
functionality!)

Additionally, added better comments to `webvtt_cue_validate_set_settings`,
and removed currently nonsensical "return if status is PARSE_ERROR or BAD_CUE",
because n'either situation is currently possible. Will re-add a similar statement
if necessary in the near future.

a WARNING_AT macro is now provided, in order to emit a warning without returning
a value regardless of what the error handler does.